achevement

+++ app/Http/Controllers/UserController.php
@@ -10,16 +10,18 @@ use Validator;
 use ImageIntervention;
 use Storage;
 use App\Image;
+use App\Http\Requests\UserValidation;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     public function index()
     {
+        $this->authorize('admin');
         $roles = Role::all();
         $positions = Position::all();
         $users = User::all();
-        return view('editUsers/user',compact('users','positions','roles'));
+        return view('editUsers/user', compact('users', 'positions', 'roles'));
     }
 
     public function indexMembre()
@@ -27,38 +29,32 @@ class UserController extends Controller
         $roles = Role::all();
         $positions = Position::all();
         $users = User::all();
-        return view('editUsers/membre',compact('users','positions','roles'));
+        return view('editUsers/membre', compact('users', 'positions', 'roles'));
     }
 
-    public function create(Request $request)
+    public function create(UserValidation $request)
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required',
-            'email' => 'required',
-            'password'=>'required',
-            'role'=>'required',
-            // 'position'=>'required',
-        ]);
+        $this->authorize('admin');
 
         if ($validator->fails()) {
             return redirect()->back()
-                        ->withErrors($validator)
-                        ->withInput();
+                ->withErrors($validator)
+                ->withInput();
         }
 
         $users = new User;
         $users->name = $request->name;
         $users->email = $request->email;
-        $users->password = Hash::make($request->password);
+        if ($request->password) {
+            $users->password = Hash::make($request->password);
+        }
         $users->roles_id = $request->role;
 
         if ($request->newPos == null && $request->position != 1) {
             $users->positions_id = $request->position;
-        }
-        elseif ($request->newPos == null) {
-            $users->positions_id = NULL;
-        }
-        else {
+        } elseif ($request->newPos == null) {
+            $users->positions_id = null;
+        } else {
             $position = new Position;
             $position->name = $request->newPos;
 
@@ -68,50 +64,54 @@ class UserController extends Controller
         }
 
 
-        if ($request->image == true) {
-            $image = $request->image;
+        // if ($request->image == true) {
+        $image = $request->image;
 
-            $filename = time().$image->hashName();
+        $filename = time() . $image->hashName();
 
             //resize
-            $redimension = ImageIntervention::make($image)->resize(360, 448)->save();
-            Storage::put('public/img/redimensionner/'.$filename,$redimension);
+        $redimension = ImageIntervention::make($image)->resize(360, 448)->save();
+        Storage::put('public/img/redimensionner/' . $filename, $redimension);
+        $imageUsers = ImageIntervention::make($image)->resize(90, 90)->save();
+        Storage::put('public/img/imageUsers/' . $filename, $imageUsers);
 
             //envoi DB
-            $table = new Image;
-            $table->url = $filename;
+        $table = new Image;
+        $table->url = $filename;
 
-            $table->save();
-        }
+        $table->save();
+        // }
 
         $users->image_user = $table->id;
 
         $users->save();
-        return redirect()->back()->with('success','User create !');
+        return redirect()->back()->with('success', 'User create !');
     }
 
     public function edit(User $users, $id)
     {
-        $users=User::find($id);
-        $roles=Role::all();
-        $positions=Position::all();
-        return view('editUsers/editUser', compact('users','roles','positions'));
+        $this->authorize('admin');
+        $users = User::find($id);
+        $roles = Role::all();
+        $positions = Position::all();
+        return view('editUsers/editUser', compact('users', 'roles', 'positions'));
     }
 
-    public function update(Request $request, $id)
+    public function update(UserValidation $request, $id)
     {
-        $users=User::find($id);
+        $users = User::find($id);
         $roles = Role::get();
         $positions = Position::get();
         $users->name = $request->name;
         $users->email = $request->email;
-        $users->password = $request->password;
+        $users->password = Hash::make($request->password);;
         $users->roles_id = $request->role;
 
-        if ($request->newPos == null) {
+        if ($request->newPos == null && $request->position != 1) {
             $users->positions_id = $request->position;
-        }
-        else {
+        } elseif ($request->newPos == null) {
+            $users->positions_id = null;
+        } else {
             $position = new Position;
             $position->name = $request->newPos;
 
@@ -119,31 +119,38 @@ class UserController extends Controller
 
             $users->positions_id = $position->id;
         }
+
         if ($request->image == true) {
             $image = $request->image;
 
-            $filename = time().$image->hashName();
+            $filename = time() . $image->hashName();
 
             //resize
             $redimension = ImageIntervention::make($image)->resize(360, 448)->save();
-            Storage::put('public/img/redimensionner/'.$filename,$redimension);
+            Storage::put('public/img/redimensionner/' . $filename, $redimension);
+            $imageUsers = ImageIntervention::make($image)->resize(90, 90)->save();
+            Storage::put('public/img/imageUsers/' . $filename, $imageUsers);
 
             //envoi DB
             $table = new Image;
             $table->url = $filename;
 
             $table->save();
+
+            $users->image_user = $table->id;
         }
+
         $users->save();
-        return redirect('/user')->with('success','User update !');
+        return redirect('/user')->with('success', 'User update !');
     }
 
     public function destroy(User $users, $id)
     {
-        $users=User::find($id);
+        $this->authorize('admin');
+        $users = User::find($id);
         $users->delete();
 
-        return redirect()->back()->with('success','User delete !');
+        return redirect()->back()->with('success', 'User delete !');
     }
